created packages for src files, and added the Inputdialog class
Please create a new package for your code, then if you need access to a class import it from the package. This will reduce naming errors, help separate the code we worked on, and make managing classes easier. Also, we will be running the input through inputOutput.runInput(), there is an example there